+++ packages/google-vertex/src/google-vertex-image-settings.ts
@@ -1,5 +1,6 @@
 export type GoogleVertexImageModelId =
   | 'imagen-3.0-generate-001'
+  | 'imagen-3.0-generate-002'
   | 'imagen-3.0-fast-generate-001'
   | (string & {});