+++ packages/google/src/google-generative-ai-settings.ts
@@ -1,7 +1,6 @@
 export type GoogleGenerativeAIModelId =
   // Stable models
   // https://ai.google.dev/gemini-api/docs/models/gemini
-  | 'gemini-2.0-flash-001'
   | 'gemini-1.5-flash'
   | 'gemini-1.5-flash-latest'
   | 'gemini-1.5-flash-001'
@@ -13,13 +12,18 @@ export type GoogleGenerativeAIModelId =
   | 'gemini-1.5-pro-latest'
   | 'gemini-1.5-pro-001'
   | 'gemini-1.5-pro-002'
-  // Experimental models
-  // https://ai.google.dev/gemini-api/docs/models/experimental-models
-  | 'gemini-2.5-pro-exp-03-25'
-  | 'gemini-2.0-flash-lite-preview-02-05'
+  | 'gemini-2.0-flash'
+  | 'gemini-2.0-flash-001'
+  | 'gemini-2.0-flash-live-001'
+  | 'gemini-2.0-flash-lite'
   | 'gemini-2.0-pro-exp-02-05'
   | 'gemini-2.0-flash-thinking-exp-01-21'
   | 'gemini-2.0-flash-exp'
+  // Experimental models
+  // https://ai.google.dev/gemini-api/docs/models/experimental-models
+  | 'gemini-2.5-pro-exp-03-25'
+  | 'gemini-2.5-pro-preview-05-06'
+  | 'gemini-2.5-flash-preview-04-17'
   | 'gemini-exp-1206'
   | 'gemma-3-27b-it'
   | 'learnlm-1.5-pro-experimental'