+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,89 @@
 
 ### Patch Changes
 
+- Updated dependencies [ed0ebeb]
+  - ai@4.3.16
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.3.15
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [a295521]
+  - ai@4.3.14
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.3.13
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [1ed3755]
+- Updated dependencies [46cb332]
+  - ai@4.3.12
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.3.11
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0432959]
+  - ai@4.3.10
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [b69a253]
+  - ai@4.3.9
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [6e8a73b]
+  - ai@4.3.8
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [f4f3945]
+  - ai@4.3.7
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [beef951]
+- Updated dependencies [bd41167]
+  - ai@4.3.6
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [452bf12]
+  - ai@4.3.5
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [013faa8]
   - ai@4.3.4