+++ packages/cerebras/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @ai-sdk/cerebras
 
+## 0.2.14
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+  - @ai-sdk/openai-compatible@0.2.14
+
+## 0.2.13
+
+### Patch Changes
+
+- Updated dependencies [23571c9]
+  - @ai-sdk/openai-compatible@0.2.13
+
+## 0.2.12
+
+### Patch Changes
+
+- Updated dependencies [13492fe]
+  - @ai-sdk/openai-compatible@0.2.12
+
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [b5c9cd4]
+  - @ai-sdk/openai-compatible@0.2.11
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [beef951]
+  - @ai-sdk/provider@1.1.3
+  - @ai-sdk/openai-compatible@0.2.10
+  - @ai-sdk/provider-utils@2.2.7
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [1bbc698]
+  - @ai-sdk/openai-compatible@0.2.9
+
 ## 0.2.8
 
 ### Patch Changes