+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/cohere
 
+## 1.2.10
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+
+## 1.2.9
+
+### Patch Changes
+
+- 033f445: fix (provider/cohere): tool calling
+
+## 1.2.8
+
+### Patch Changes
+
+- Updated dependencies [beef951]
+  - @ai-sdk/provider@1.1.3
+  - @ai-sdk/provider-utils@2.2.7
+
 ## 1.2.7
 
 ### Patch Changes