+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,88 @@
 # @ai-sdk/valibot
 
+## 0.1.28
+
+### Patch Changes
+
+- Updated dependencies [ed0ebeb]
+  - ai@4.3.16
+
+## 0.1.27
+
+### Patch Changes
+
+- ai@4.3.15
+
+## 0.1.26
+
+### Patch Changes
+
+- Updated dependencies [a295521]
+  - ai@4.3.14
+
+## 0.1.25
+
+### Patch Changes
+
+- ai@4.3.13
+
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies [1ed3755]
+- Updated dependencies [46cb332]
+  - ai@4.3.12
+
+## 0.1.23
+
+### Patch Changes
+
+- ai@4.3.11
+
+## 0.1.22
+
+### Patch Changes
+
+- Updated dependencies [0432959]
+  - ai@4.3.10
+
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [b69a253]
+  - ai@4.3.9
+
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [6e8a73b]
+  - ai@4.3.8
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [f4f3945]
+  - ai@4.3.7
+
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [beef951]
+- Updated dependencies [bd41167]
+  - ai@4.3.6
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [452bf12]
+  - ai@4.3.5
+
 ## 0.1.16
 
 ### Patch Changes